Update to Confluent ZooKeeper and Kafka 5.5.3
Update ZooKeeper and Kafka to version 5.5.3 of the Confluent distribution.

Update Traefik configuration for latest 2.3 version

Traefik has been updated to version 2.3 along with the associated
configuration changes to the services it routes traffic to.

The services have been updated to use the nip.io site to resolve
their hostnames back to the machine this deployment is running on.
This means that no changes are needed to the /etc/hosts file and Traefik
is able to route the requests based on the hostname header.

- Fix the Kafka transporter connection string as this uses the Kafka bootstrap
url and not the ZooKeeper connect string in later versions of kafka-node.
- Linting configuration and cleaned up JS services, README, and Docker Compose file
- Add default logging options to Docker logging driver in Docker Compose file
- Add internal and external broker listeners for communication within the
Docker Compose network for the deployment, as well as communication from the Docker host
- Add JMX port configuration to access metrics
- Add example environment variables for Docker Compose deployment
- Removed unnecessary "links" entries from Docker Compose file
- Removed "depends_on" entries from some services as the usage of Kafka and Redis
is optional and those containers may not need to be started.

The default service broker configuration is provided in the moleculer.config.js file,
this is overridden by the environment variables when running the Docker deployment.