Define user routes within user service and whitelist
The routes for user operations are declared within the Users service and whitelisted within
the API Gateway. The action handlers do not follow the standard default REST handler names
as they are provided with the username in the path which is used to resolve the actual
database identifier. The default REST action handler names are provided by the DBService
mixin, hence needing to separate these and explicitly declare for the REST HTTP verbs.

Without whitelisting these the usage of the autoAliases setting with a value of true
means that all action handlers were exposed, including the default ones from the database
mixin. This meant that GET /users/:id was calling the database service with 'users.get' and not
'users.getUser' as was expected for the GET /users/:username route