Add support for multiple Kafka consumers
Support has been added to the Kafka service mixin to allow multiple
Kafka consumers to be created by a single service. This provides
the means for services to be able to perform operations based on
events the can consume from multiple topics.

Update services to use Kafka consumer object to supply properties