Run multistage Docker build and use Google Distroless image
Use the Google Distroless image as the final image in the build stages to
reduce image size and more secure.

Node has been downgraded to version 14 to match the local version and prevent the
lock file from updating when building within the image. The package-lock.json
file was not previously being copied into the build stage.