Associate inventory items with an order
Items are reserved based on their product and each item, based on id,
that has been reserved is added to the order. The "populates" functionality
is used so that when an order is accessed via get/list/find then the inventory
items are populated into the document and not just the array of ids.

Remove redundant "fields" entries as they should be within the "settings" section, but aren't being use to filter out fields anyway.

Use the context to make calls to other services for nested calls and tracing