Use NODE_ID as environment variable for unique identifier
The environment variable NODEID is reserved by Moleculer so was overriding the generated
unique identifier that contained the hostname. This would lead to multiple nodes with the same
identifier when scaled up within Docker. The environment variable within the config and docker-compose
is now NODE_ID which is used as a prefix to provide the final node identifier.